requirements.txt
alto-xml==0.0.5
astroid==2.8.6
autopep8==1.6.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.0
charset-normalizer==2.1.1
cryptography==36.0.1
distlib==0.3.4
filelock==3.5.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
img2pdf==0.4.3
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.7.1
mypy==0.910
mypy-extensions==0.4.3
oauth2client==4.1.3
pdf2image==1.16.0
pdfminer.six==20211012
pipenv==2022.1.8
platformdirs==2.4.1
protobuf==4.21.12
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pylint==2.11.1
pylint-quotes==0.2.3
pyparsing==3.0.9
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tesserocr==2.5.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.13.1
virtualenv-clone==0.5.7
wrapt==1.13.3
youtube3==1.2.2
